+++ example/lib/app/routes/app_navigation.dart
@@ -1,9 +1,8 @@
 
 
-
-
-import 'package:example/app/routes/app_routes.dart';
 import 'package:get/get.dart';
+
+import 'app_routes.dart';
 
 class AppNavigation {
   static void getBackWithValue(dynamic result) {
@@ -14,10 +13,32 @@ class AppNavigation {
     Get.back();
   }
 
-  static void pushToDashboard() {
-    Get.offAllNamed(AppRoutes.dashboard);
+
+  static void pushToDashboard({String? email, String? password}) {
+    Get.offAllNamed(
+      AppRoutes.dashboard,
+      arguments: {
+        'email': email,
+        'password': password,
+      },
+    );
+  }
+  static void pushToLogin() {
+    Get.offAllNamed(AppRoutes.login);
   }
 
+  static void pushToRegistration() {
+    Get.offAllNamed(AppRoutes.registration);
+  }
+
+  // static void getBackWithValue(dynamic result) {
+  //   Get.back(result: result);
+  // }
+  //
+  // static void getBack() {
+  //   Get.back();
+  // }
+  //
   // static void pushToLogin() {
   //   Get.toNamed(AppRoutes.login);
   // }