+++ example/lib/app/services/auth_service.dart
@@ -1,13 +1,9 @@
-
-import 'package:ub_code_structure/app/api/end_point.dart';
-import 'package:ub_code_structure/app/api/web_service_helper.dart';
-import 'package:ub_code_structure/app/models/product_model.dart';
-import 'package:ub_code_structure/app/utils/utils.dart';
-
-
+import '../api/end_point.dart';
+import '../api/web_service_helper.dart';
+import '../models/product_model.dart';
+import '../utils/utils.dart';
 
 class AuthService {
-
   Future<List<ProductModel>?> getDetails() async {
     final webserviceHelper = WebserviceHelper(endPointType: Endpoint.posts, path: []);
     final map = await webserviceHelper.postNormal();
@@ -25,7 +21,7 @@ class AuthService {
       params: {
         "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
         "body":
-        "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto",
+            "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto",
       },
     );
     final map = await webserviceHelper.postNormal();
@@ -35,12 +31,8 @@ class AuthService {
     return products;
   }
 
-  Future<ProductModel?> putDetails()async{
-    final webserviceHelper = WebserviceHelper(
-      endPointType: Endpoint.postsId,
-      path: ['2'],
-
-    );
+  Future<ProductModel?> putDetails() async {
+    final webserviceHelper = WebserviceHelper(endPointType: Endpoint.postsId, path: ['2']);
 
     final map = await webserviceHelper.postNormal();
     Utils.debugLog(title: "putDetails", object: " $map ${map['data'].runtimeType}");